Update to allow conditional removal of the lock

+++ tests/package-runner.sh
@@ -1,12 +1,17 @@
 #!/bin/bash
 
+# Set defaults.
+CLEAN_LOCK=0 # Used to trigger if composer.lock and the vendor folder should be deleted.
+PACKAGE='' # Used to store the package name passed in the args.
+
+# This function fires if no package name is passed in the command line and will run all packages.
 function run_packages_tests {
 	echo "Running tests for all packages:"
 	export PACKAGES='./packages/**'
-	for PACKAGE in $PACKAGES
+	for P in $PACKAGES
 	do
-		if [ -d "$PACKAGE" ]; then
-			folder="$(basename $PACKAGE)"
+		if [ -d "$P" ]; then
+			folder="$(basename $P)"
 			result="run_package_test $folder"
 
 			if $result; then
@@ -19,14 +24,17 @@ function run_packages_tests {
 	done
 }
 
+# This function runs a particular package.
 function run_package_test {
 	package=$1;
 
 	if [ -f "./packages/$package/phpunit.xml.dist" ]; then
 			echo "Running tests for package \`$package\` "
 			cd "./packages/$package"
-			rm composer.lock
-			rm -rf vendor
+			if [ 1 == $CLEAN_LOCK ]; then
+				rm composer.lock
+				rm -rf vendor
+			fi
 			composer install
 			cd "../.."
 			result="yarn docker:compose exec wordpress phpunit --configuration=/var/www/html/wp-content/plugins/jetpack/packages/$package/phpunit.xml.dist"
@@ -42,9 +50,24 @@ function run_package_test {
 
 }
 
+for arg in "$@"
+do
+	case $arg in
+		-c|--clean-lock)
+			CLEAN_LOCK=1
+			shift
+			;;
+		*)
+			PACKAGE="$1"
+			shift
+			;;
+	esac
+done
 
-if [ -f "./packages/$1/phpunit.xml.dist"  ]; then
-	run_package_test "$1"
+
+
+if [ -f "./packages/$PACKAGE/phpunit.xml.dist"  ]; then
+	run_package_test "$PACKAGE"
 else
 	echo "You did not specify a package. Running all of them can take a good amount of time."
 	echo "Do you wish to run tests for all packages?"
